+++ Source/Core/Core/HW/GCMemcard/GCMemcardDirectory.h
@@ -12,6 +12,7 @@
 #include "Common/Event.h"
 #include "Core/HW/GCMemcard/GCIFile.h"
 #include "Core/HW/GCMemcard/GCMemcard.h"
+#include "Core/HW/GCMemcard/GCMemcardBase.h"
 
 // Uncomment this to write the system data of the memorycard from directory to disc
 //#define _WRITE_MC_HEADER 1
