+++ Source/Core/Core/HW/GCMemcard/GCMemcardRaw.h
@@ -11,6 +11,7 @@
 #include "Common/Event.h"
 #include "Common/Flag.h"
 #include "Core/HW/GCMemcard/GCMemcard.h"
+#include "Core/HW/GCMemcard/GCMemcardBase.h"
 
 class PointerWrap;
